Fix for crash when initial setting text in view
If text view is created programmatically, no font is defined and
a text string with BB code is set the app crashed in because of
missing font attributes. The fix is using the system font for this
edge case now. The example app is also using the formatted string
accordingly.